requirements.txt
python-telegram-bot>=21.0
openai>=1.30.0
google-auth>=2.28.0
google-api-python-client>=2.120.0
pytz>=2024.1
Pillow>=10.0.0
